Move the API endpoints to a better location
